Move out v1/v2 handling.
