+++ src/libfsm/dfavm.c
@@ -20,10 +20,6 @@
 
 #include "dfavm.h"
 
-#define DEBUG_ENCODING     0
-#define DEBUG_VM_EXECUTION 0
-#define DEBUG_VM_OPCODES   0
-
 // VM state:
 //
 //   Conceptually, the VM is composed of:
@@ -220,79 +216,6 @@
 //   DD=11 is reserved for future use
 //
 
-#define DFAVM_VARENC_MAJOR 0x00
-#define DFAVM_VARENC_MINOR 0x01
-
-#define DFAVM_FIXEDENC_MAJOR 0x00
-#define DFAVM_FIXEDENC_MINOR 0x02
-
-#define DFAVM_MAGIC "DFAVM$"
-
-static enum dfavm_io_result
-dfavm_v1_save(FILE *f, const struct dfavm_v1 *vm)
-{
-	/* write eight byte header */
-	unsigned char header[8];
-	unsigned char ilen[4];
-
-	memcpy(&header[0], DFAVM_MAGIC, 6);
-	header[6] = DFAVM_VARENC_MAJOR;
-	header[7] = DFAVM_VARENC_MINOR;
-
-	if (fwrite(&header, sizeof header, 1, f) != 1) {
-		return -1;
-	}
-
-	/* write out instructions */
-	ilen[0] = (vm->len >>  0) & 0xff;
-	ilen[1] = (vm->len >>  8) & 0xff;
-	ilen[2] = (vm->len >> 16) & 0xff;
-	ilen[3] = (vm->len >> 24) & 0xff;
-
-	if (fwrite(&ilen[0],sizeof ilen, 1, f) != 1) {
-		return -1;
-	}
-
-	if (fwrite(vm->ops, 1, vm->len, f) != vm->len) {
-		return -1;
-	}
-
-	return 0;
-}
-
-static enum dfavm_io_result
-dfavm_load_v1(FILE *f, struct dfavm_v1 *vm)
-{
-	unsigned char ilen[4];
-	uint32_t len;
-	unsigned char *instr;
-	size_t nread;
-
-	/* read number of instructions */
-	if (fread(&ilen[0], sizeof ilen, 1, f) != 1) {
-		return DFAVM_IO_ERROR_READING;
-	}
-
-	len = ((uint32_t)ilen[0] << 0) | ((uint32_t)ilen[1] << 8) |
-		((uint32_t)ilen[1] << 16) | ((uint32_t)ilen[1] << 24);
-
-	instr = malloc(len);
-	if (instr == NULL) {
-		return DFAVM_IO_OUT_OF_MEMORY;
-	}
-
-	nread = fread(instr, 1, len, f);
-	if (nread != len) {
-		free(instr);
-		return (nread > 0) ? DFAVM_IO_SHORT_READ : DFAVM_IO_ERROR_READING;
-	}
-
-	vm->ops = instr;
-	vm->len = len;
-
-	return DFAVM_IO_OK;
-}
-
 enum dfavm_io_result
 fsm_dfavm_save(FILE *f, const struct fsm_dfavm *vm)
 {
@@ -327,8 +250,9 @@ fsm_dfavm_load(FILE *f, struct fsm_dfavm *vm)
 	return DFAVM_IO_UNSUPPORTED_VERSION;
 }
 
-static const char *
-cmp_name(int cmp) {
+const char *
+cmp_name(int cmp)
+{
 	switch (cmp) {
 	case VM_CMP_ALWAYS: return "";   break;
 	case VM_CMP_LT:     return "LT"; break;
@@ -1505,272 +1429,6 @@ assign_rel_dests(struct dfavm_assembler *a)
 	}
 }
 
-static struct fsm_dfavm *
-encode_opasm_v1(const struct dfavm_assembler *a)
-{
-	static const struct fsm_dfavm zero;
-
-	struct fsm_dfavm *ret;
-	struct dfavm_v1 *vm;
-	size_t total_bytes;
-	size_t i;
-	size_t off;
-	unsigned char *enc;
-
-	ret = malloc(sizeof *ret);
-	*ret = zero;
-
-	ret->version_major = DFAVM_VARENC_MAJOR;
-	ret->version_minor = DFAVM_VARENC_MINOR;
-
-	vm = &ret->u.v1;
-
-	total_bytes = a->nbytes;
-
-	enc = malloc(total_bytes);
-
-	vm->ops = enc;
-	vm->len = total_bytes;
-
-	for (off = 0, i = 0; i < a->ninstr; i++) {
-		unsigned char bytes[6];
-		unsigned char cmp_bits, instr_bits, rest_bits;
-		const struct dfavm_vm_op *op = &a->instr[i];
-		int nb = 1;
-
-		cmp_bits   = 0;
-		instr_bits = 0;
-		rest_bits  = 0;
-
-		cmp_bits = op->cmp;
-		if (cmp_bits > 7) {
-			goto error;
-		}
-
-		if (op->cmp != VM_CMP_ALWAYS) {
-			bytes[1] = op->cmp_arg;
-			nb++;
-		}
-
-		switch (op->instr) {
-		case VM_OP_STOP:
-			instr_bits = 0x0;
-			rest_bits  = (op->u.stop.end_bits == VM_END_SUCC) ? 0x1 : 0x0;
-			break;
-
-		case VM_OP_FETCH:
-			instr_bits = 0x1;
-			rest_bits  = (op->u.fetch.end_bits == VM_END_SUCC) ? 0x1 : 0x0;
-			break;
-
-		case VM_OP_BRANCH:
-			instr_bits = 0x2;
-			switch (op->u.br.dest) {
-			case VM_DEST_SHORT:
-				{
-					int8_t dst;
-
-					assert(op->u.br.rel_dest >= INT8_MIN);
-					assert(op->u.br.rel_dest <= INT8_MAX);
-
-					rest_bits = 0;
-
-				       	dst = op->u.br.rel_dest;
-					memcpy(&bytes[nb], &dst, sizeof dst);
-					nb += sizeof dst;
-				}
-				break;
-
-			case VM_DEST_NEAR:
-				{
-					int16_t dst;
-
-					assert(op->u.br.rel_dest >= INT16_MIN);
-					assert(op->u.br.rel_dest <= INT16_MAX);
-
-					rest_bits = 1;
-
-				       	dst = op->u.br.rel_dest;
-					memcpy(&bytes[nb], &dst, sizeof dst);
-					nb += sizeof dst;
-				}
-				break;
-
-			case VM_DEST_FAR:
-				{
-					int32_t dst;
-
-					assert(op->u.br.rel_dest >= INT32_MIN);
-					assert(op->u.br.rel_dest <= INT32_MAX);
-
-					rest_bits = 2;
-
-				       	dst = op->u.br.rel_dest;
-					memcpy(&bytes[nb], &dst, sizeof dst);
-					nb += sizeof dst;
-				}
-				break;
-
-			case VM_DEST_NONE:
-			default:
-				goto error;
-			}
-
-			break;
-
-		default:
-			goto error;
-		}
-
-		bytes[0] = (cmp_bits << 5) | (instr_bits << 3) | (rest_bits);
-#if DEBUG_ENCODING
-		fprintf(stderr, "enc[%4zu] instr=0x%02x cmp=0x%02x rest=0x%02x b=0x%02x\n",
-			off, instr_bits, cmp_bits, rest_bits, bytes[0]);
-#endif /* DEBUG_ENCODING */
-
-		memcpy(&enc[off], bytes, nb);
-		off += nb;
-
-		assert(off <= total_bytes);
-	}
-
-	assert(off == total_bytes);
-
-	return ret;
-
-error:
-	/* XXX - cleanup */
-	return NULL;
-}
-
-static struct fsm_dfavm *
-encode_opasm_v2(const struct dfavm_assembler *a)
-{
-	static const struct fsm_dfavm zero;
-
-	struct fsm_dfavm *ret;
-	struct dfavm_v2 *vm;
-	size_t total_instr;
-	size_t i;
-	uint32_t *enc;
-	uint32_t *abuf;
-	uint32_t alen,acap;
-
-	ret = malloc(sizeof *ret);
-	*ret = zero;
-
-	ret->version_major = DFAVM_FIXEDENC_MAJOR;
-	ret->version_minor = DFAVM_FIXEDENC_MINOR;
-
-	vm = &ret->u.v2;
-
-	total_instr = a->ninstr;
-	enc = malloc(total_instr * sizeof *enc);
-
-	vm->ops = enc;
-	vm->len = total_instr;
-
-	abuf = NULL;
-	alen = 0;
-	acap = 0;
-
-	for (i = 0; i < a->ninstr; i++) {
-		unsigned char cmp_bits, instr_bits, result_bit, cmp_arg;
-		const struct dfavm_vm_op *op = &a->instr[i];
-		uint32_t instr;
-		uint16_t dest_arg;
-
-		switch (op->instr) {
-		case VM_OP_STOP:
-			instr_bits = VM_V2_OP_STOP;
-			cmp_bits   = op->cmp;
-			cmp_arg    = op->cmp_arg;
-			result_bit = (op->u.stop.end_bits == VM_END_SUCC) ? 0x1 : 0x0;
-			dest_arg   = 0;
-			break;
-
-		case VM_OP_FETCH:
-			instr_bits = VM_V2_OP_FETCH;
-			cmp_bits   = 0;
-			cmp_arg    = 0;
-			result_bit = (op->u.fetch.end_bits == VM_END_SUCC) ? 0x1 : 0x0;
-			dest_arg   = 0;
-			break;
-
-		case VM_OP_BRANCH:
-			{
-				uint32_t dest_state = op->u.br.dest_index;
-				int64_t state_diff = (int64_t)dest_state - (int64_t)i;
-
-				if (state_diff >= INT16_MIN && state_diff <= INT16_MAX) {
-					instr_bits = VM_V2_OP_BRANCH;
-					dest_arg = (int16_t)state_diff;
-				}
-				else {
-					/* allocate address in address table */
-					instr_bits = VM_V2_OP_IBRANCH;
-
-					if (alen >= acap) {
-						size_t new_acap;
-						uint32_t *new_abuf;
-
-						if (acap < 16) {
-							new_acap = 16;
-						} else if (acap < 1024) {
-							new_acap = acap * 2;
-						} else {
-							new_acap = acap + acap/2;
-						}
-
-						new_abuf = realloc(abuf, new_acap * sizeof abuf[0]);
-						if (new_abuf == NULL) {
-							goto error;
-						}
-
-						abuf = new_abuf;
-						acap = new_acap;
-
-						vm->abuf = new_abuf;
-					}
-
-					assert(alen < acap);
-					assert(abuf != NULL);
-
-					abuf[alen] = dest_state;
-					dest_arg = alen++;
-					vm->alen = alen;
-				}
-
-				cmp_bits   = op->cmp;
-				cmp_arg    = op->cmp_arg;
-				result_bit = 0;
-			}
-			break;
-		}
-
-		if (cmp_bits > 7) {
-			goto error;
-		}
-
-		instr = (((uint32_t)instr_bits) << 28) | (((uint32_t)cmp_bits)<<25) | (((uint32_t)result_bit)<<24) |
-			(((uint32_t)cmp_arg)<<16) | dest_arg;
-
-#if DEBUG_ENCODING
-		fprintf(stderr, "enc[%4zu] instr=0x%02x cmp=0x%02x result=0x%02x cmp_arg=0x%02x dest_arg=0x%04x (%d | %u) instr=0x%08lx\n",
-			off, instr_bits, cmp_bits, result_bit, (unsigned)cmp_arg, (unsigned)dest_arg, (int16_t)dest_arg, (unsigned)dest_arg,
-			(unsigned long)instr);
-#endif /* DEBUG_ENCODING */
-
-		enc[i] = instr;
-	}
-
-	return ret;
-
-error:
-	/* XXX - cleanup */
-	return NULL;
-}
-
 static void
 dump_states(FILE *f, struct dfavm_assembler *a)
 {
@@ -1938,376 +1596,6 @@ void
 fsm_vm_free(struct fsm_dfavm *vm)
 {
 	(void)vm;
-}
-
-static uint32_t
-running_print_op_v1(const unsigned char *ops, uint32_t pc, const char *sp, const char *buf, size_t n, char ch, FILE *f) {
-	int op, cmp, end, dest;
-	unsigned char b = ops[pc];
-	op = (b >> 3) & 0x03;
-	cmp = (b>>5)  & 0x07;
-	dest = b & 0x03;
-	end  = b & 0x01;
-
-	fprintf(f, "[%4lu sp=%zd n=%zu ch=%3u '%c' end=%d] ",
-		(unsigned long)pc, sp-buf, n, (unsigned char)ch, isprint(ch) ? ch : ' ', end);
-
-	switch (op) {
-	case VM_OP_FETCH:
-		fprintf(f, "FETCH%s\n", (end == VM_END_FAIL) ? "F" : "S");
-		break;
-
-	case VM_OP_STOP:
-		fprintf(f, "STOP%s%s", (end == VM_END_FAIL) ? "F" : "S", cmp_name(cmp));
-		if (cmp != VM_CMP_ALWAYS) {
-			int arg = ops[++pc];
-			if (isprint(arg)) {
-				fprintf(f, " '%c'", arg);
-			} else {
-				fprintf(f, " %d", arg);
-			}
-		}
-		fprintf(f, "\n");
-		break;
-
-	case VM_OP_BRANCH:
-		fprintf(f, "BR%s%s", (dest == 0 ? "S" : (dest == 1 ? "N" : "F")), cmp_name(cmp));
-		if (cmp != VM_CMP_ALWAYS) {
-			int arg = ops[++pc];
-			if (isprint(arg)) {
-				fprintf(f, " '%c',", arg);
-			} else {
-				fprintf(f, " %d,", arg);
-			}
-		}
-
-
-		{
-			int32_t rel;
-			unsigned char bytes[4] = { 0, 0, 0, 0 };
-
-			if (dest == 0) {
-				union {
-					uint8_t u;
-					int8_t  i;
-				} packed;
-
-				packed.u = ops[pc+1];
-				bytes[0] = packed.u;
-				rel = packed.i;
-				pc += 1;
-			} else if (dest == 1) {
-				union {
-					uint16_t u;
-					int16_t  i;
-				} packed;
-
-				packed.u = (ops[pc+1] | (ops[pc+2] << 8));
-				bytes[0] = ops[pc+1];
-				bytes[1] = ops[pc+2];
-				rel = packed.i;
-				pc += 2;
-			} else {
-				union {
-					uint32_t u;
-					int32_t  i;
-				} packed;
-				packed.u = ops[pc+1] | (ops[pc+2] << 8) | (ops[pc+3] << 16) | (ops[pc+4] << 24);
-				bytes[0] = ops[pc+1];
-				bytes[1] = ops[pc+2];
-				bytes[2] = ops[pc+3];
-				bytes[3] = ops[pc+4];
-				rel = packed.i;
-				pc += 4;
-			}
-
-			fprintf(f, " %ld 0x%02x 0x%02x 0x%02x 0x%02x\n",
-				(long)rel, bytes[0], bytes[1], bytes[2], bytes[3]);
-		}
-		break;
-
-	default:
-		fprintf(f, "UNK\n");
-	}
-
-	return pc;
-}
-
-static enum dfavm_state
-vm_match_v1(const struct dfavm_v1 *vm, struct vm_state *st, const char *buf, size_t n)
-{
-	const char *sp, *last;
-	int ch;
-
-	if (st->state != VM_MATCHING) {
-		return st->state;
-	}
-
-	ch   = 0;
-	sp   = buf;
-	last = buf + n;
-
-	for (;;) {
-		int op;
-		unsigned char b;
-
-		/* decode instruction */
-		b = vm->ops[st->pc];
-		op = (b>>3)&0x03;
-
-		(void)running_print_op_v1;  /* make clang happy */
-#if DEBUG_VM_EXECUTION
-		running_print_op_v1(vm->ops, st->pc, sp, buf, n, ch, stderr);
-#endif /* DEBUG_VM_EXECUTION */
-
-		if (op == VM_OP_FETCH) {
-			int end;
-
-			end = b & 0x01;
-			if (sp >= last) {
-				st->fetch_state = end;
-				return VM_MATCHING;
-			}
-
-			ch = (unsigned char) *sp++;
-			st->pc++;
-		} else {
-			int cmp, end, dest, dest_nbytes;
-			int result;
-			uint32_t off;
-
-			/* either STOP or BRANCH */
-			off = st->pc+1;
-			cmp  = b >> 5;
-
-			result = 0;
-			if (cmp == VM_CMP_ALWAYS) {
-				result = 1;
-			} else {
-				int arg;
-
-				arg = vm->ops[off++];
-
-				switch (cmp) {
-				case VM_CMP_LT: result = ch <  arg; break;
-				case VM_CMP_LE: result = ch <= arg; break;
-				case VM_CMP_GE: result = ch >= arg; break;
-				case VM_CMP_GT: result = ch >  arg; break;
-				case VM_CMP_EQ: result = ch == arg; break;
-				case VM_CMP_NE: result = ch != arg; break;
-				}
-			}
-
-			dest = b & 0x3;
-			dest_nbytes = 1 << dest;
-
-			if (result) {
-				int32_t rel;
-
-				if (op == VM_OP_STOP) {
-					end = b & 0x01;
-					st->state = (end == VM_END_FAIL) ? VM_FAIL : VM_SUCCESS;
-					return st->state;
-				}
-
-				if (dest == 0) {
-					rel = (int8_t)(vm->ops[off++]);
-				} else if (dest == 1) {
-					union {
-						uint16_t u;
-						int16_t  i;
-					} packed;
-
-					packed.u = (vm->ops[off+0] | (vm->ops[off+1] << 8));
-					rel = packed.i;
-					off += 2;
-				} else {
-					union {
-						uint32_t u;
-						int32_t  i;
-					} packed;
-					packed.u = vm->ops[off+0] | (vm->ops[off+1] << 8) | (vm->ops[off+2] << 16) | (vm->ops[off+3] << 24);
-					rel = packed.i;
-				}
-
-				st->pc += rel;
-			} else if (op == VM_OP_BRANCH) {
-				st->pc = off + dest_nbytes;
-			} else {
-				st->pc = off;
-			}
-		}
-	}
-
-	return VM_FAIL;
-}
-
-#define V2DEC(instr,op,cmp,end,cmp_arg,dest_arg) do {  \
-	uint32_t tmp ## __LINE__ = (instr);            \
-	(op)       = (tmp ## __LINE__) >> 28;          \
-	(cmp)      = ((tmp ## __LINE__) >> 25) & 0x7;  \
-	(end)      = ((tmp ## __LINE__) >> 24) & 0x1;  \
-	(cmp_arg)  = ((tmp ## __LINE__) >> 16) & 0xff; \
-	(dest_arg) = (tmp ## __LINE__) & 0xffff;       \
-} while(0)
-
-static void
-running_print_op_v2(const struct dfavm_v2 *vm, uint32_t pc, const char *sp, const char *buf, size_t n, char ch, FILE *f)
-{
-	int op, cmp, end, arg, dest;
-
-	V2DEC(vm->ops[pc], op,cmp,end, arg, dest);
-
-	fprintf(f, "[%4lu sp=%zd n=%zu ch=%3u '%c' end=%d] ",
-		(unsigned long)pc, sp-buf, n, (unsigned char)ch, isprint(ch) ? ch : ' ', end);
-
-	switch (op) {
-	case VM_V2_OP_FETCH:
-		fprintf(f, "FETCH%s\n", (end == VM_END_FAIL) ? "F" : "S");
-		break;
-
-	case VM_V2_OP_STOP:
-		fprintf(f, "STOP%s%s", (end == VM_END_FAIL) ? "F" : "S", cmp_name(cmp));
-		if (cmp != VM_CMP_ALWAYS) {
-			if (isprint(arg)) {
-				fprintf(f, " '%c'", arg);
-			} else {
-				fprintf(f, " %d", arg);
-			}
-		}
-		fprintf(f, "\n");
-		break;
-
-	case VM_V2_OP_BRANCH:
-		fprintf(f, "BR%s", cmp_name(cmp));
-		if (cmp != VM_CMP_ALWAYS) {
-			if (isprint(arg)) {
-				fprintf(f, " '%c',", arg);
-			} else {
-				fprintf(f, " %d,", arg);
-			}
-		}
-
-		{
-			union {
-				uint16_t u;
-				int16_t  rel;
-			} packed;
-
-			packed.u = dest;
-
-			fprintf(f, " %d %u\n", (int)packed.rel, dest);
-		}
-		break;
-
-	case VM_V2_OP_IBRANCH:
-		fprintf(f, "IBR%s", cmp_name(cmp));
-		if (cmp != VM_CMP_ALWAYS) {
-			if (isprint(arg)) {
-				fprintf(f, " '%c',", arg);
-			} else {
-				fprintf(f, " %d,", arg);
-			}
-		}
-
-		{
-			assert(dest >= 0 && (uint32_t)dest < vm->alen);
-			uint32_t dest_addr = vm->abuf[dest];
-			fprintf(f, " dest=%lu index=%d\n", (unsigned long)dest_addr, dest);
-		}
-		break;
-
-	default:
-		fprintf(f, "UNK[op=%d cmp=%d end=%d arg=%d dest=%d]\n", op,cmp,end,arg, dest);
-		break;
-	}
-}
-
-static enum dfavm_state
-vm_match_v2(const struct dfavm_v2 *vm, struct vm_state *st, const char *buf, size_t n)
-{
-	const char *sp, *last;
-	int ch;
-
-	if (st->state != VM_MATCHING) {
-		return st->state;
-	}
-
-	ch   = 0;
-	sp   = buf;
-	last = buf + n;
-
-	for (;;) {
-		int op, cmp, end, arg, dest;
-
-		assert(st->pc < vm->len);
-
-		/* decode instruction */
-		V2DEC(vm->ops[st->pc], op,cmp,end, arg, dest);
-
-		(void)running_print_op_v2;  /* make clang happy */
-#if DEBUG_VM_EXECUTION
-		running_print_op_v2(vm, st->pc, sp, buf, n, ch, stderr);
-#endif /* DEBUG_VM_EXECUTION */
-
-		if (op == VM_V2_OP_FETCH) {
-			if (sp >= last) {
-				st->fetch_state = end;
-				return VM_MATCHING;
-			}
-
-			ch = (unsigned char) *sp++;
-			st->pc++;
-		} else {
-			int result;
-
-			/* either STOP or BRANCH */
-			result = 1;
-			if (cmp != VM_CMP_ALWAYS) {
-				switch (cmp) {
-				case VM_CMP_LT: result = ch <  arg; break;
-				case VM_CMP_LE: result = ch <= arg; break;
-				case VM_CMP_GE: result = ch >= arg; break;
-				case VM_CMP_GT: result = ch >  arg; break;
-				case VM_CMP_EQ: result = ch == arg; break;
-				case VM_CMP_NE: result = ch != arg; break;
-				}
-			}
-
-			if (result) {
-				switch (op) {
-				case VM_V2_OP_STOP:
-					st->state = (end == VM_END_FAIL) ? VM_FAIL : VM_SUCCESS;
-					return st->state;
-
-				case VM_V2_OP_BRANCH:
-					{
-						union {
-							uint16_t u;
-							int16_t  rel;
-						} packed;
-
-						packed.u = dest;
-						st->pc += packed.rel;
-					}
-					break;
-
-				case VM_V2_OP_IBRANCH:
-					assert((uint32_t)dest < vm->alen);
-					st->pc = vm->abuf[dest];
-					break;
-
-				default:
-					// should not reach!
-					abort();
-				}
-			} else {
-				st->pc++;
-			}
-		}
-	}
-
-	return VM_FAIL;
 }
 
 static enum dfavm_state
